requirements.txt
psycopg2-binary==2.9.9
groq>=0.11.0
python-dotenv==1.0.0
dataclasses>=0.6; python_version<"3.7"